+++ CHANGELOG.md
@@ -2,9 +2,14 @@
 
 ## sys-info-extended
 
+### v0.9.1
+
+- Added some documentation for the functions that implemented later than fork.
+- `is_program_installed()` functions algorithm improved.
+
 ### v0.9.0
 
-- Added `UserConfigurations` struct, that includes home directory and shell preference of the user and `get_home_dir_and_shell()` function for getting that struct.
+- Added `UserConfigurations` struct for linux, that includes home directory and shell preference of the user and `get_home_dir_and_shell()` function for getting that struct.
 - cc and libc versions upgraded.
 
 ### v0.8.1